requirements.txt
alembic==1.13.3
altgraph==0.17.4
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
APScheduler==3.10.4
billiard==4.2.1
blinker==1.8.2
bsdiff4==1.2.5
cachelib==0.13.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
Deprecated==1.2.15
dsdev-utils @ file:///C:\Users\rcagt\OneDrive\Documents\radcom22\RADCOM\finalradcom\Radcomtest\ds\dsdev-utils-1.3.0
et-xmlfile==1.1.0
fastapi==0.115.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.5
MarkupSafe==3.0.2
msgspec==0.18.6
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pbr==6.1.0
pefile==2023.2.7
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyNaCl==1.5.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyUpdater==4.0
pywin32-ctypes==0.2.3
redis==5.2.0
requests==2.32.3
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
stevedore==3.5.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.20
uvicorn==0.31.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.45.0
wrapt==1.16.0
